==> 2-Descriptor_calculation.ipynb <==
## Descriptor calculation

import pandas as pd

from rdkit.Chem import PandasTools
from rdkit_utils import FP
from rdkit.Chem import Descriptors

import pickle

df = pd.read_excel('mol_inchi_for_modelling.xlsx')

PandasTools.AddMoleculeColumnToFrame(df, smilesCol='usmiles', molCol='mol')

### Calculate 6 usual descriptors 

df['logp'] = df['mol'].apply(Descriptors.MolLogP)
df['mwt'] = df['mol'].apply(Descriptors.MolWt)
df['hbd'] = df['mol'].apply(Descriptors.NumHDonors)
df['hba'] = df['mol'].apply(Descriptors.NumHAcceptors)
df['rtb'] = df['mol'].apply(Descriptors.NumRotatableBonds)
df['tpsa'] = df['mol'].apply(Descriptors.TPSA)

### Calculate fingerprint

df['FP'] = df['mol'].apply(FP)

### Export 

pickle.dump(df, open('mols.pkl', 'wb'))
