==> 4-Job_submission.ipynb <==
## Submit training jobs
### To be more efficient, the model training run on our CPU cluster
This scripts provides an example

#### Notebook meant to be exported as .py

### Arguments generated before are required in addition to script shown in notebook 5
eg: python3.5 4-Job_submission_comparison.py /nfs/panda/chembl/nbosc/input_files /nfs/panda/chembl/nbosc/comp_results RF

import os
from sys import argv

if __name__ == '__main__':
    ### input_dir is where are located the 
    path, input_dir, output_dir, method = argv
    
    if not os.path.exists(output_dir):
        os.makedirs(output_dir)
        
    for each_dir in [item for item in os.listdir(input_dir) if not 'runtimes.csv' in item]:
        df_prot_filename = input_dir+'/'+each_dir+'/'+each_dir+'.pkl'

        command = 'bsub -M 10000 -R "rusage[mem=10000]" -n 6 -o job_output_comp python3.5 5-Modeling_QSAR_and_CP.py {} {} {} {}'.format(df_prot_filename, output_dir, each_dir, method)
        #print(command)
        os.system(command)
